feat: Реализовать период воздержания и потраченные/сэкономленные средства

+++ featureDashboard/impl/src/main/java/ru/health/featuredashboard/impl/presentation/dashboard/DashboardUiState.kt
@@ -8,13 +8,6 @@ internal data class DashboardUiState(
     val hasNotifications: Boolean = false,
     val health: Int = 100,
     val abstinenceDuration: Duration = Duration.ZERO,
-    val savedMoney: Float = 0f
+    val savedMoney: Double = 0.0
 )
-
-//internal fun DashboardUiState.copy(info: DashboardInfo) = DashboardUiState(
-//    hasNotifications = info.hasNotifications,
-//    health = info.health,
-//    lastRelapsedDate = info.abstinenceDuration.time,
-//    savedMoney = info.savedMoney
-//)
