+++ featureDashboard/impl/src/main/java/ru/health/featuredashboard/impl/presentation/dashboard/ui/banner/BannerFeed.kt
@@ -24,7 +24,7 @@ import kotlin.math.roundToInt
 @Composable
 internal fun BannerFeed(
     modifier: Modifier = Modifier,
-    maxPrice: Float
+    maxPrice: Double
 ) {
     val windowInfo = LocalWindowInfo.current
     val density = LocalDensity.current
